+++ client/src/App.jsx
@@ -18,6 +18,7 @@ import ReferralBoard from './pages/ReferralBoard';
 import PublicProfile from './pages/PublicProfile';
 import ProfileSettings from './pages/ProfileSettings';
 import Reviews from './pages/Reviews';
+import Bookmarks from './pages/Bookmarks';
 import ProtectedRoute from './components/ProtectedRoute';
 
 function App() {
@@ -39,6 +40,7 @@ function App() {
           <Route path="/alumni" element={<ProtectedRoute><AlumniDirectory /></ProtectedRoute>} />
           <Route path="/referrals" element={<ProtectedRoute><ReferralBoard /></ProtectedRoute>} />
           <Route path="/reviews" element={<ProtectedRoute><Reviews /></ProtectedRoute>} />
+          <Route path="/bookmarks" element={<ProtectedRoute><Bookmarks /></ProtectedRoute>} />
           <Route path="/moderation" element={<ProtectedRoute><ModerationQueue /></ProtectedRoute>} />
           <Route path="/chats" element={<ProtectedRoute><Chats /></ProtectedRoute>} />
           <Route path="/chat/:userId" element={<ProtectedRoute><ChatThread /></ProtectedRoute>} />